requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cleanup==8.0.0
django-froala-editor==4.1.0
Pillow==10.0.0
sqlparse==0.4.4
tzdata==2023.3
